move search-related methods from (Abstract)MetadataRepository into SearchService and SearchMapper
because that's where they belong (the repository interface should be independent of search)
